Fix 7 Dart compilation errors blocking iOS App Store archive build
Agent-Logs-Url: https://github.com/zedraxa/y2notes2/sessions/104690e8-8a0c-4175-9332-7b3e6c231ae4

+++ lib/features/effects/engine/easing_curves.dart
@@ -101,5 +101,5 @@ abstract class EasingCurves {
 
   /// Exponential ease-out: fast start, decelerating to zero velocity.
   static double expoOut(double t) =>
-      t >= 1 ? 1 : 1 - math.pow(2, -10 * t);
+      t >= 1 ? 1.0 : 1.0 - math.pow(2, -10 * t).toDouble();
 }
